+++ packages/terra-core-docs/src/terra-dev-site/test/button-group/ButtonGroupText.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ButtonGroup from '../../../ButtonGroup';
+import ButtonGroup from 'terra-button-group';
 
 const TextButtonGroup = () => (
   <ButtonGroup id="button-group-text">