+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
-* Added
+* Changed
   * Migrate doc from terra-button.
+
+* Changed
+  * Migrate doc from terra-button-group.
 
 ## 1.11.0 - (October 8, 2021)